pyproject.toml
[project]
name = "rede_social"
version = "0.1.0"
description = ""
authors = [
    {name = "Joao Gabriel Iuzviak Mantagute"}
]
readme = "README.md"
packages = [{include = "*", from="src"}]
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.0,<0.117.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dev = "fastapi dev dev.py"
prod = "uvicorn src.api.application:app --host 0.0.0.0 --port 8000 --reload"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"

